+++ crates/nu-plugin/README.md
@@ -20,24 +20,4 @@ The steps are as follows:
 ## Benchmark
 Here is a simple benchmark for different protocol for encoding/decoding nushell table, with different rows and columns.  You can simply run `cargo bench` to run benchmark.
 
-### If you want to gather and analysis benchmark data...
-1. It's recommended to use cargo-criterion to run benchmark for future analysing:
-```
-cargo install cargo-criterion
-```
-
-2. then run the following commands to generate benchmark result:
-```
-cargo criterion --message-format=json | save "result.json"
-```
-
-3. filter noisy data from cargo-criterion:
-```
-let aa = (open result.json | lines | where (not ($it | str contains "group-complete")))
-("[" + ($aa | str collect ',') + "]") | save -r result.json
-```
-
-4. open `result.json` file and do anything you want to gather benchmark result, e.g:
-```
-open result.json | select id mean | flatten mean
-```
+The relative html report is in `target/criterion/report/index.html`.
